+++ CMakeLists.txt
@@ -10,8 +10,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 file(GLOB src_files
 	"src/*.h"
-	"src/*.cpp"
-	"src/test/*.cpp"
+	"src/main.cpp"
 )
 
 if(WIN32)
