+++ src/headers/lexer.h
@@ -8,7 +8,3 @@ struct Lexer
     u32 row;
     std::vector<Token> tokens;
 };
-
-
-// NOTE: return -> was there an error?
-bool tokenize(Lexer &lexer,const std::vector<std::string> &file);
